Extract SystemProperties into plugin-toolkit module
Allows reduction of boilerplate related to system properties
