+++ lib/trino-plugin-toolkit/src/main/java/io/trino/plugin/base/util/SystemProperties.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.kudu;
+package io.trino.plugin.base.util;
 
 import java.util.Objects;
